git-svn-id: svn://svn.chromium.org/chrome/branches/chrome_official_branch/src@2472 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/disk_cache/disk_cache_test_base.cc
@@ -1,31 +1,6 @@
-// Copyright 2008, Google Inc.
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are
-// met:
-//
-//    * Redistributions of source code must retain the above copyright
-// notice, this list of conditions and the following disclaimer.
-//    * Redistributions in binary form must reproduce the above
-// copyright notice, this list of conditions and the following disclaimer
-// in the documentation and/or other materials provided with the
-// distribution.
-//    * Neither the name of Google Inc. nor the names of its
-// contributors may be used to endorse or promote products derived from
-// this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+// Copyright (c) 2006-2008 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
 #include "net/disk_cache/disk_cache_test_base.h"
 
@@ -33,7 +8,7 @@
 #include "net/disk_cache/disk_cache_test_util.h"
 #include "net/disk_cache/mem_backend_impl.h"
 
-void DiskCacheTestBase::SetMaxSize(int size) {
+void DiskCacheTestWithCache::SetMaxSize(int size) {
   size_ = size;
   if (cache_impl_)
     EXPECT_TRUE(cache_impl_->SetMaxSize(size));
@@ -42,7 +17,7 @@ void DiskCacheTestBase::SetMaxSize(int size) {
     EXPECT_TRUE(mem_cache_->SetMaxSize(size));
 }
 
-void DiskCacheTestBase::InitCache() {
+void DiskCacheTestWithCache::InitCache() {
   if (mask_)
     implementation_ = true;
 
@@ -52,10 +27,11 @@ void DiskCacheTestBase::InitCache() {
     InitDiskCache();
 
   ASSERT_TRUE(NULL != cache_);
-  ASSERT_EQ(0, cache_->GetEntryCount());
+  if (first_cleanup_)
+    ASSERT_EQ(0, cache_->GetEntryCount());
 }
 
-void DiskCacheTestBase::InitMemoryCache() {
+void DiskCacheTestWithCache::InitMemoryCache() {
   if (!implementation_) {
     cache_ = disk_cache::CreateInMemoryCacheBackend(size_);
     return;
@@ -71,9 +47,10 @@ void DiskCacheTestBase::InitMemoryCache() {
   ASSERT_TRUE(mem_cache_->Init());
 }
 
-void DiskCacheTestBase::InitDiskCache() {
+void DiskCacheTestWithCache::InitDiskCache() {
   std::wstring path = GetCachePath();
-  ASSERT_TRUE(DeleteCache(path.c_str()));
+  if (first_cleanup_)
+    ASSERT_TRUE(DeleteCache(path.c_str()));
 
   if (!implementation_) {
     cache_ = disk_cache::CreateCacheBackend(path, force_creation_, size_);
@@ -95,17 +72,19 @@ void DiskCacheTestBase::InitDiskCache() {
 }
 
 
-void DiskCacheTestBase::TearDown() {
+void DiskCacheTestWithCache::TearDown() {
   delete cache_;
 
   if (!memory_only_) {
     std::wstring path = GetCachePath();
     EXPECT_TRUE(CheckCacheIntegrity(path));
   }
+
+  PlatformTest::TearDown();
 }
 
 // We are expected to leak memory when simulating crashes.
-void DiskCacheTestBase::SimulateCrash() {
+void DiskCacheTestWithCache::SimulateCrash() {
   ASSERT_TRUE(implementation_ && !memory_only_);
   cache_impl_->ClearRefCountForTest();
 
@@ -124,3 +103,9 @@ void DiskCacheTestBase::SimulateCrash() {
     cache_impl_->SetMaxSize(size_);
   ASSERT_TRUE(cache_impl_->Init());
 }
+
+void DiskCacheTestWithCache::SetTestMode() {
+  ASSERT_TRUE(implementation_ && !memory_only_);
+  cache_impl_->SetUnitTestMode();
+}
+
